requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
future==0.18.2
idna==2.10
iso8601==0.1.14
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
pandas==1.3.0
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
tqdm==4.61.2
urllib3==1.26.6
Werkzeug==2.0.1
